segundo commmit?

+++ src/main/java/Auto.java
@@ -1,11 +1,11 @@
 public class Auto {
 
-    private int Velocidad;
+    private double Velocidad;
     public Auto(){
         this.Velocidad = 50;
     }
 
-    public int getVelocidad() {
+    public double getVelocidad() {
         return this.Velocidad;
     }
